# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Tutor API",
    "version": "1.0.0",
    "description": "AI-powered learning platform backend API.\n\n**Modules:** Auth, Organizations, Academic Structure, Materials, Lessons, Flashcards, Quizzes, Chat (RAG), Progress, Analytics, Search, Notifications, Usage, Jobs\n\n**Response format:** `{ success, message, data }`"
  },
  "servers": [{ "url": "http://localhost:7001", "description": "Development" }],
...
  "paths": {
...
    "/organizations": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Create a new organization",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Springfield High School" },
                  "logo": { "type": "string", "format": "uri" },
                  "subscriptionPlan": {
                    "type": "string",
                    "enum": ["FREE", "BASIC", "PRO", "ENTERPRISE"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Organization" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Organizations"],
        "summary": "List organizations",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Organizations retrieved" }
        }
      }
    },
    "/organizations/{id}": {
      "get": {
        "tags": ["Organizations"],
        "summary": "Get organization by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Organization retrieved" },
          "404": { "description": "Organization not found" }
        }
      },
      "patch": {
        "tags": ["Organizations"],
        "summary": "Update organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "logo": { "type": "string", "format": "uri" },
                  "subscriptionPlan": {
                    "type": "string",
                    "enum": ["FREE", "BASIC", "PRO", "ENTERPRISE"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Organization updated" }
        }
      },
      "delete": {
        "tags": ["Organizations"],
        "summary": "Soft delete organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Organization deleted" }
        }
      }
    },
    "/organizations/{id}/members": {
      "get": {
        "tags": ["Organizations"],
        "summary": "List organization members",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Members retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/OrganizationMember" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Organizations"],
        "summary": "Add member to organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "role"],
                "properties": {
                  "userId": { "type": "string" },
                  "role": {
                    "type": "string",
                    "enum": ["TEACHER", "STUDENT", "PARENT"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Member added" }
        }
      }
    },
    "/organizations/{id}/members/{userId}": {
      "delete": {
        "tags": ["Organizations"],
        "summary": "Remove member from organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Member removed" }
        }
      }
    },
    "/organizations/{id}/members/{userId}/suspend": {
      "patch": {
        "tags": ["Organizations"],
        "summary": "Suspend organization member",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Member suspended" }
        }
      }
    },
    "/organizations/{id}/invites": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Invite member by email",
        "description": "Sends an invitation email with a token link. Requires SCHOOL_ADMIN or SUPER_ADMIN.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "role": { "type": "string", "enum": ["TEACHER", "STUDENT", "PARENT"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Invitation sent" }
        }
      }
    },
    "/organizations/invites/accept": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Accept organization invitation",
        "description": "Public endpoint. Creates account if new user, or links existing account.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": { "type": "string" },
                  "firstName": { "type": "string", "description": "Required for new accounts" },
                  "lastName": { "type": "string", "description": "Required for new accounts" },
                  "password": { "type": "string", "minLength": 8, "description": "Required for new accounts" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Invitation accepted" },
          "400": { "description": "Invalid or expired token" }
        }
      }
    },
    "/organizations/{id}/parent-links": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Link parent to student",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["parentId", "studentId"],
                "properties": {
                  "parentId": { "type": "string" },
                  "studentId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Parent linked to student" }
        }
      }
    },
    "/organizations/{id}/assignments/teacher": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Assign teacher to subject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["subjectId", "teacherId"],
                "properties": {
                  "subjectId": { "type": "string" },
                  "teacherId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Teacher assigned" }
        }
      }
    },
    "/organizations/{id}/assignments/student": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Enroll student in subject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["subjectId", "studentId"],
                "properties": {
                  "subjectId": { "type": "string" },
                  "studentId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Student enrolled" }
        }
      }
    },
...
    "/analytics/organization/{organizationId}": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Organization analytics",
        "description": "Active students, materials, AI usage, lesson completion. Requires SCHOOL_ADMIN.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "organizationId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Organization analytics retrieved" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT access token from /auth/login"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Success" },
          "data": { "type": "object" }
        }
      },
...
      "Organization": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "logo": { "type": "string" },
          "subscriptionPlan": {
            "type": "string",
            "enum": ["FREE", "BASIC", "PRO", "ENTERPRISE"]
          },
          "ownerId": { "type": "string" },
          "teachers": { "type": "array", "items": { "type": "string" } },
          "students": { "type": "array", "items": { "type": "string" } },
          "parents": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "OrganizationMember": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["SUPER_ADMIN", "SCHOOL_ADMIN", "TEACHER", "PARENT", "STUDENT"] },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "SUSPENDED", "DELETED"] },
          "avatar": { "type": "string" },
          "isOwner": { "type": "boolean" }
        }
      },
...
    }
  }
}